requirements.txt
Flask==3.0.3
Flask_And_Redis==1.0.0
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.30
Werkzeug==3.0.3
